# specs/SPEC/api-spec.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "InfinityMix API",
    "version": "1.0.0",
    "description": "API for InfinityMix, an AI-powered mashup generation platform."
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "API v1"
    }
  ],
...
  "paths": {
...
    "/users/me": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Get current user's profile information",
        "operationId": "getCurrentUser",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Current user profile",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                },
                "example": {
                  "id": "a1b2c3d4-e5f6-7890-1234-567890abcdef",
                  "email": "user@example.com",
                  "username": "MashupMaster",
                  "created_at": "2025-11-28T10:00:00Z",
                  "updated_at": "2025-11-28T10:00:00Z"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "put": {
        "tags": [
          "users"
        ],
        "summary": "Update current user's profile information",
        "operationId": "updateCurrentUser",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserRequest"
              },
              "example": {
                "username": "NewMashupDJ",
                "email": "new_email@example.com"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User profile updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                },
                "example": {
                  "id": "a1b2c3d4-e5f6-7890-1234-567890abcdef",
                  "email": "new_email@example.com",
                  "username": "NewMashupDJ",
                  "created_at": "2025-11-28T10:00:00Z",
                  "updated_at": "2025-11-28T11:30:00Z"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "delete": {
        "tags": [
          "users"
        ],
        "summary": "Delete current user's account",
        "operationId": "deleteCurrentUser",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteAccountRequest"
              },
              "example": {
                "password": "StrongPassword123!"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "User account deleted successfully"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
...
    "/mashups": {
      "get": {
        "tags": [
          "mashups"
        ],
        "summary": "Get a list of all mashups generated by the current user",
        "operationId": "listUserMashups",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Page number for pagination",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Number of items per page",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of user's mashups",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MashupListResponse"
                },
                "example": {
                  "page": 1,
                  "limit": 10,
                  "total": 2,
                  "data": [
                    {
                      "id": "m1a2s3h4-u5p6-7890-1234-567890abcdef",
                      "user_id": "a1b2c3d4-e5f6-7890-1234-567890abcdef",
                      "duration_seconds": 120,
                      "status": "completed",
                      "output_path": "https://storage.example.com/mashups/m1a2s3h4-u5p6.mp3",
                      "output_format": "mp3",
                      "generation_time_ms": 75000,
                      "created_at": "2025-11-28T12:05:00Z",
                      "updated_at": "2025-11-28T12:06:15Z"
                    },
                    {
                      "id": "b2c3d4e5-f6a7-8901-2345-67890abcdef0",
                      "user_id": "a1b2c3d4-e5f6-7890-1234-567890abcdef",
                      "duration_seconds": 60,
                      "status": "completed",
                      "output_path": "https://storage.example.com/mashups/b2c3d4e5-f6a7.mp3",
                      "output_format": "mp3",
                      "generation_time_ms": 45000,
                      "created_at": "2025-11-28T13:00:00Z",
                      "updated_at": "2025-11-28T13:00:45Z"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "UserResponse": {
        "type": "object",
        "required": [
          "id",
          "email",
          "username",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Unique identifier for the user."
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "User's email address."
          },
          "username": {
            "type": "string",
            "description": "User's display name."
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "Timestamp when the account was created."
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "description": "Timestamp when the account was last updated."
          }
        }
      },
      "UpdateUserRequest": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "minLength": 3,
            "maxLength": 50,
            "description": "New username for the user."
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "New email address for the user."
          }
        },
        "minProperties": 1,
        "description": "Fields to update for the user profile. At least one field is required."
      },
      "DeleteAccountRequest": {
        "type": "object",
        "required": [
          "password"
        ],
        "properties": {
          "password": {
            "type": "string",
            "description": "Current password for account deletion confirmation."
          }
        }
      },
...
      "MashupResponse": {
        "type": "object",
        "required": [
          "id",
          "user_id",
          "duration_seconds",
          "status",
          "output_format",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Unique identifier for the mashup."
          },
          "user_id": {
            "type": "string",
            "format": "uuid",
            "description": "ID of the user who generated the mashup."
          },
          "duration_seconds": {
            "type": "integer",
            "description": "Target duration of the mashup in seconds."
          },
          "status": {
            "type": "string",
            "enum": [
              "queued",
              "processing",
              "completed",
              "failed"
            ],
            "description": "Current status of the mashup generation."
          },
          "output_path": {
            "type": "string",
            "format": "uri",
            "nullable": true,
            "description": "URL to the generated mashup file (available if status is 'completed')."
          },
          "output_format": {
            "type": "string",
            "description": "Format of the output file (e.g., 'mp3')."
          },
          "generation_time_ms": {
            "type": "integer",
            "nullable": true,
            "description": "Time taken for generation in milliseconds (available if status is 'completed')."
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "Timestamp when the generation request was made."
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "description": "Timestamp when the mashup status was last updated."
          }
        }
      },
      "MashupListResponse": {
        "type": "object",
        "required": [
          "page",
          "limit",
          "total",
          "data"
        ],
        "properties": {
          "page": {
            "type": "integer",
            "description": "Current page number."
          },
          "limit": {
            "type": "integer",
            "description": "Number of items per page."
          },
          "total": {
            "type": "integer",
            "description": "Total number of mashups available."
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MashupResponse"
            }
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "required": [
          "message"
        ],
        "properties": {
          "message": {
            "type": "string",
            "description": "A human-readable error message."
          },
          "code": {
            "type": "string",
            "description": "An optional, application-specific error code."
          }
        }
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Invalid input provided (e.g., missing fields, invalid format).",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "message": "Invalid email format."
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Authentication required or invalid credentials.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "message": "Invalid email or password."
            }
          }
        }
      },
...
      "NotFound": {
        "description": "The requested resource was not found.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "message": "Mashup with ID 'xyz' not found."
            }
          }
        }
      },
      "Conflict": {
        "description": "Resource already exists (e.g., email already registered).",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "message": "Email 'user@example.com' is already registered."
            }
          }
        }
      },
      "InternalServerError": {
        "description": "An unexpected server error occurred.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "message": "An internal server error occurred. Please try again later."
            }
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT Authorization header using the Bearer scheme."
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}